requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0.post0
pytz==2024.1
gunicorn==23.0.0